fixes small SNAFUs, updates doc

+++ provision-drill.sh
@@ -17,14 +17,17 @@ set -e # exit on error immediately
 ###############################################################################
 # Global variables
 
-SCRIPT_PATH=`dirname $0`
+DRILL_TARGET=/opt/drill/apache-drill-0.8.0/bin/drillbit.sh
 
-echo Provisioning Apache Drill 0.8
+## Install Apache Drill if not yet done
+if [[ ! -f $DRILL_TARGET ]]; then 
+    echo "Drill seems not to be set up, yet. Going to install it now ..."
+    wget http://getdrill.org/drill/download/apache-drill-0.8.0.tar.gz
+    sudo mkdir -p /opt/drill
+    sudo tar -xvzf apache-drill-0.8.0.tar.gz -C /opt/drill
+    rm apache-drill-0.8.0.tar.gz
+fi
 
-## Install Drill
-wget http://getdrill.org/drill/download/apache-drill-0.8.0.tar.gz
-sudo mkdir -p /opt/drill
-sudo tar -xvzf apache-drill-0.8.0.tar.gz -C /opt/drill
-rm apache-drill-0.8.0.tar.gz
+echo "Drill is installed and ready to rock."
 
 exit 0
